abstracts findMatchingPoolDisk into helper function
